back-end/warning/alert/active/requirements.txt
mysql-connector-python
paho-mqtt
requests